LayoutTests/svg/dynamic-updates/SVGTextElement-svgdom-rotate-prop.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="resources/SVGTestCase.js"></script>
<script src="../../resources/js-test-pre.js"></script>
<script src="../../fast/repaint/resources/repaint.js"></script>
</head>
<body onload="runRepaintTest()">
<h1>SVG 1.1 dynamic update tests</h1>
<p id="description"></p>
<div id="console"></div>
<script>
// [Name] SVGTextElement-svgdom-rotate-prop.js
// [Expected rendering result] 'Test passed' message - and a series of PASS messages
description("Tests dynamic updates of the 'rotate' property of the SVGTextElement object")
createSVGTestCase();

var textElement = createSVGElement("text");
textElement.setAttribute("x", "0");
textElement.setAttribute("y", "215");
textElement.setAttribute("rotate", "-90");
textElement.textContent = "Text content";
rootSVGElement.appendChild(textElement);

shouldBe("textElement.rotate.baseVal.getItem(0).value", "-90");

function repaintTest() {
    textElement.rotate.baseVal.getItem(0).value = 0;
    shouldBe("textElement.rotate.baseVal.getItem(0).value", "0");

    completeTest();
}

var successfullyParsed = true;
</script>
</body>
</html>
