third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGTextElement-svgdom-rotate-prop-expected.txt

SVG 1.1 dynamic update tests
Text content

Tests dynamic updates of the 'rotate' property of the SVGTextElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textElement.rotate.baseVal.getItem(0).value is -90
PASS textElement.rotate.baseVal.getItem(0).value is 0
PASS successfullyParsed is true

TEST COMPLETE
